Move Scene creation tests to separate module
